requirements.txt
streamlit==1.29.0
pandas==2.1.1
numpy==1.26.0
python-docx==0.8.11
PyPDF2==3.0.1
spacy==3.7.2
python-pptx==0.6.21
scikit-learn==1.3.0
plotly==5.17.0
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
google-generativeai==0.3.1
python-dotenv==1.0.0
